Fixes2

+++ app/src/main/java/com/example/photolink/MainActivity.kt
@@ -1,7 +1,8 @@
 package com.example.photolink
 
+import android.Manifest
 import android.app.usage.UsageEvents
-import android.graphics.BitmapFactory
+import android.content.pm.PackageManager
 import android.net.Uri
 import android.os.Bundle
 import android.util.Log
@@ -10,11 +11,14 @@ import android.view.MenuItem
 import android.widget.Toast
 import androidx.appcompat.app.AlertDialog
 import androidx.appcompat.app.AppCompatActivity
+import androidx.core.app.ActivityCompat
+import androidx.core.content.ContextCompat
 import androidx.core.net.toFile
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentManager
 import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProvider
+import com.example.photolink.CameraFragment.Companion.REQUEST_CODE_PERMISSIONS
 import com.example.photolink.Model.IteamPlace
 import com.example.photolink.api.RequestApiImpl
 import com.example.photolink.api.RequestRepository
@@ -22,11 +26,18 @@ import io.reactivex.android.schedulers.AndroidSchedulers
 import io.reactivex.disposables.CompositeDisposable
 import io.reactivex.schedulers.Schedulers
 import kotlinx.android.synthetic.main.activity_main.*
-import kotlinx.android.synthetic.main.fragment_description.*
-import okhttp3.ResponseBody
 import java.io.File
+import java.io.FileInputStream
+import java.io.FileOutputStream
+import java.io.IOException
+
 
 class MainActivity : AppCompatActivity(), PlaceInteractor, RowInteractor, CameraInteractor, DescriptionInteractor, ServerSettingsInteractor {
+
+    companion object {
+        private val FILE_NAME = "content.txt"
+        private val REQUEST_WRITE_STORAGE_REQUEST_CODE = 45
+    }
 
     private val newsRepository by lazy(UsageEvents.Event.NONE) { RequestRepository(RequestApiImpl(this)) }
     private val compositeDisposable = CompositeDisposable()
@@ -58,12 +69,20 @@ class MainActivity : AppCompatActivity(), PlaceInteractor, RowInteractor, Camera
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
         //makeCurrentFragment(mainFragment)
+        requestAppPermissions()
+        if (openText() != null) {
+            if (hasReadPermissions()) {
+                mainViewModel.setBaseURI(openText()!!)
+            }
+        }
         mainViewModel.baseURI.observe(this, Observer {
+            if (hasWritePermissions()) {
+                saveText(it)
+            }
             newsRepository.updateURI(it)
             loadPlace()
         })
     }
-
 
     fun loadPlace() {
         val disposable = newsRepository.loadListGson().subscribeOn(Schedulers.io())
@@ -75,7 +94,6 @@ class MainActivity : AppCompatActivity(), PlaceInteractor, RowInteractor, Camera
                     makeCurrentFragment(startFragment)
                 }, {
                     AlertDialog.Builder(this).setMessage("Ошибка загрузки").setMessage(it.message).show()
-
                 }
                 )//.updateListPost(news as MutableList<Post>)}
         compositeDisposable.add(disposable)
@@ -90,18 +108,22 @@ class MainActivity : AppCompatActivity(), PlaceInteractor, RowInteractor, Camera
 
     override fun onBackPressed() {
         var flag = true
-        if (supportFragmentManager.fragments.last() is ServerSettings) {
-            flag = false
-        }
-        super.onBackPressed()
-        if (flag) {
-            Log.d("Place-valueList", mainViewModel.lastName.value.toString())
-            if (!supportFragmentManager.fragments.isEmpty()) {
-                if (!(supportFragmentManager.fragments.last() is CameraFragment)) {
-                    mainViewModel.removePlace()
-                }
+        if (!supportFragmentManager.fragments.isEmpty()) {
+            if (supportFragmentManager.fragments.last() is ServerSettings) {
+                flag = false
             }
-            Log.d("Place-value", mainViewModel.namePalace.value)
+            super.onBackPressed()
+            if (flag) {
+                Log.d("Place-valueList", mainViewModel.lastName.value.toString())
+                if (!supportFragmentManager.fragments.isEmpty()) {
+                    if (!(supportFragmentManager.fragments.last() is CameraFragment)) {
+                        mainViewModel.removePlace()
+                    }
+                }
+                Log.d("Place-value", mainViewModel.namePalace.value)
+            }
+        } else {
+            super.onBackPressed()
         }
     }
 
@@ -116,6 +138,11 @@ class MainActivity : AppCompatActivity(), PlaceInteractor, RowInteractor, Camera
         mainViewModel.addPlace(name)
     }
 
+    fun onStateDefoltFragment() {
+        supportFragmentManager.popBackStack(null, FragmentManager.POP_BACK_STACK_INCLUSIVE);
+        mainViewModel.startMain()
+    }
+
     override fun onRefreshPlace() {
         supportFragmentManager.popBackStack(null, FragmentManager.POP_BACK_STACK_INCLUSIVE);
         loadPlace()
@@ -124,7 +151,8 @@ class MainActivity : AppCompatActivity(), PlaceInteractor, RowInteractor, Camera
     }
 
     override fun onClickRow(name: String) {
-        supportFragmentManager.beginTransaction().replace(R.id.nav_frame, CameraFragment()).addToBackStack(null).commit()
+        val fragment = CameraFragment.newInstance(name)
+        supportFragmentManager.beginTransaction().replace(R.id.nav_frame, fragment).addToBackStack(null).commit()
         mainViewModel.addPlace(name)
     }
 
@@ -140,23 +168,43 @@ class MainActivity : AppCompatActivity(), PlaceInteractor, RowInteractor, Camera
         val arrayList = ArrayList<File>()
         arrayList.add(file)
         Log.d("PlaceOut", mainViewModel.namePalace.value)
-        val disposable = newsRepository.saveFile(mainViewModel.namePalace.value!!, description, arrayList).subscribeOn(Schedulers.io())
-                .observeOn(AndroidSchedulers.mainThread())
-                .subscribe({
-                    val str = it.string()
-                    Log.d("Body", str)
-                    if (str.contains("file uploaded successfully")) {
-                        Toast.makeText(this, "Фото успешно загруженно", Toast.LENGTH_SHORT).show()
-                    } else {
-                        AlertDialog.Builder(this).setTitle("Ошибка закгрузки").setMessage(str).show()
+        requestAppPermissions()
+        if (hasReadPermissions() && hasWritePermissions()) {
+            val disposable = newsRepository.saveFile(mainViewModel.namePalace.value!!, description, arrayList).subscribeOn(Schedulers.io())
+                    .observeOn(AndroidSchedulers.mainThread())
+                    .subscribe({
+                        val str = it.string()
+                        Log.d("Body", str)
+                        if (str.contains("file uploaded successfully")) {
+                            Toast.makeText(this, "Фото успешно загруженно", Toast.LENGTH_SHORT).show()
+                        } else {
+                            AlertDialog.Builder(this).setTitle("Ошибка закгрузки").setMessage(str).show()
+                        }
+                        file.delete()
+                    }, {
+                        AlertDialog.Builder(this).setTitle("Ошибка закгрузки").setMessage(it.message).show()
+                        file.delete()
+                    })
+            compositeDisposable.add(disposable)
+            onBackPressed()
+        } else {
+            AlertDialog.Builder(this).setTitle("Ошибка доступа").setMessage("Нужен доступ к файлам").show()
+        }
+    }
+
+    override fun onRequestPermissionsResult(requestCode: Int, permissions: Array<out String>, grantResults: IntArray) {
+        super.onRequestPermissionsResult(requestCode, permissions, grantResults)
+        if (requestCode == REQUEST_CODE_PERMISSIONS) {
+            if (supportFragmentManager.fragments.isNotEmpty()) {
+                if (supportFragmentManager.fragments.last() is CameraFragment) {
+                    if (grantResults[0] == PackageManager.PERMISSION_GRANTED) {
+                        (supportFragmentManager.fragments.last() as CameraFragment).startCamera()
                     }
-                    file.delete()
-                }, {
-                    AlertDialog.Builder(this).setTitle("Ошибка закгрузки").setMessage(it.message).show()
-                    file.delete()
-                })
-        compositeDisposable.add(disposable)
-        onBackPressed()
+                    else{
+                        onBackPressed()}
+                }
+            }
+        }
     }
 
     override fun onOpenDescription(fileUri: Uri) {
@@ -169,15 +217,84 @@ class MainActivity : AppCompatActivity(), PlaceInteractor, RowInteractor, Camera
 
     }
 
+    override fun onBack() {
+        onBackPressed()
+    }
+
     override fun onStop() {
         mainViewModel.listURI.value?.forEach {
             it.toFile().delete()
         }
         compositeDisposable.clear()
+
         super.onStop()
     }
 
     override fun closeServerSettings() {
         onBackPressed()
+        supportFragmentManager.popBackStack(null, FragmentManager.POP_BACK_STACK_INCLUSIVE);
+        for (fragment in supportFragmentManager.fragments) {
+            if (fragment != null) {
+                supportFragmentManager.beginTransaction().remove(fragment).commit()
+            }
+        }
+        mainViewModel.startMain()
+    }
+
+    // открытие файла
+    fun openText(): String? {
+        var fin: FileInputStream? = null
+        var text: String? = null
+        try {
+            fin = openFileInput(FILE_NAME)
+            val bytes = ByteArray(fin.available())
+            fin.read(bytes)
+            text = String(bytes)
+        } catch (ex: IOException) {
+            Toast.makeText(this, ex.message, Toast.LENGTH_SHORT).show()
+        } finally {
+            try {
+                if (fin != null) fin.close()
+            } catch (ex: IOException) {
+                Toast.makeText(this, ex.message, Toast.LENGTH_SHORT).show()
+            }
+        }
+        return text
+    }
+
+    // сохранение файла
+    fun saveText(text: String) {
+        var fos: FileOutputStream? = null
+        try {
+            fos = openFileOutput(FILE_NAME, MODE_PRIVATE)
+            fos.write(text.toByteArray())
+        } catch (ex: IOException) {
+            Toast.makeText(this, ex.message, Toast.LENGTH_SHORT).show()
+        } finally {
+            try {
+                if (fos != null) fos.close()
+            } catch (ex: IOException) {
+                Toast.makeText(this, ex.message, Toast.LENGTH_SHORT).show()
+            }
+        }
+    }
+
+    private fun requestAppPermissions() {
+        if (hasReadPermissions() && hasWritePermissions()) {
+            return
+        }
+
+        ActivityCompat.requestPermissions(this, arrayOf(
+                Manifest.permission.READ_EXTERNAL_STORAGE,
+                Manifest.permission.WRITE_EXTERNAL_STORAGE
+        ), REQUEST_WRITE_STORAGE_REQUEST_CODE) // your request code
+    }
+
+    private fun hasReadPermissions(): Boolean {
+        return ContextCompat.checkSelfPermission(baseContext, Manifest.permission.READ_EXTERNAL_STORAGE) == PackageManager.PERMISSION_GRANTED
+    }
+
+    private fun hasWritePermissions(): Boolean {
+        return ContextCompat.checkSelfPermission(baseContext, Manifest.permission.WRITE_EXTERNAL_STORAGE) == PackageManager.PERMISSION_GRANTED
     }
 }
